# String concat allocation reduction