@pelletier 2017-06-28
Good catch. Thank you for this patch!